move licenses
