--- mteb/modalities.py
@@ -1,8 +1,0 @@
-from __future__ import annotations
-
-from typing_extensions import Literal
-
-MODALITIES = Literal[
-    "text",
-    "image",
-]
